Removed the support of webworkers + added CORS to the coin list response
