+++ src/helpers/consts.js
@@ -9,7 +9,7 @@ export const COIN_STATUSES = {
   ACTIVE: 'Active'
 };
 export const DONATIONS_ADDRESSES = {
-  LITECOIN: 'LLRKHC53wGRKh7326zQHvau1vDvQGpQqwx',
-  BITCOIN: '1Vxa4Phrn5Vtn2urHnPSWsiuYxaub9tjn',
-  ETHEREUM: '0xC620A3b5CB29Dee1532d39cdA5c7cE9dDE912CC4'
+  LITECOIN: 'LhMKQHwH76BGRuYNoFhPpdkkUDyJ9pQxNC',
+  BITCOIN: '1NPbjGGKp4B1UAFHGtsp6MHVUGNUzDmHi9',
+  ETHEREUM: '0xeae32e0b5fde43dad837b4e6133aec19eee25d0e'
 };
