--- src/redux/middlewares/worker.js
@@ -1,4 +1,0 @@
-import createWorkerMiddleware from './createWorkerMiddleware';
-const CoinsWorker = require('worker-loader?!../workers/CoinsWorker');
-
-export default createWorkerMiddleware(new CoinsWorker());
