+++ webpack-config/env-configs/local.js
@@ -6,7 +6,7 @@ const config = {
   ORIGINS: {
     COINS_HERO: 'http://local-coinshero.io:3001',
     COINS_IO: 'https://coincap.io',
-    CRYPTO_COMPARE: 'https://www.cryptocompare.com',
+    CRYPTO_COMPARE: 'https://cors-anywhere.herokuapp.com/https://www.cryptocompare.com',
     FIXER_IO: 'https://api.fixer.io',
     CHANGELLY: 'https://changelly.com'
   },
